+++ python/GafferSceneTest/__init__.py
@@ -179,6 +179,9 @@ from .DeleteRenderPassesTest import DeleteRenderPassesTest
 from .RenderPassWedgeTest import RenderPassWedgeTest
 from .RenderAdaptorTest import RenderAdaptorTest
 from .MeshTessellateTest import MeshTessellateTest
+from .RenderPassTypeAdaptorTest import RenderPassTypeAdaptorTest
+from .RenderPassShaderTest import RenderPassShaderTest
+from .RenderPassAdaptorTest import RenderPassAdaptorTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *