+++ python/GafferSceneUI/__init__.py
@@ -193,6 +193,8 @@ from . import DeleteRenderPassesUI
 from . import RenderPassWedgeUI
 from . import LightPositionToolUI
 from . import MeshTessellateUI
+from . import RenderPassTypeAdaptorUI
+from . import RenderPassShaderUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.